refactor(projectionChart): refactor dataset functions to use in cumulative chart
