+++ src/components/ProjectionChart.vue
@@ -2,7 +2,7 @@
 import { Line, mixins } from 'vue-chartjs';
 import moment from 'moment';
 import cloneDeep from 'lodash/cloneDeep';
-import { getLabels, buildChartDataSet, getColor } from '../utils/chartUtils.js';
+import { getLabels, buildChartDataSet, getColor, fillDatasetGaps } from '../utils/chartUtils.js';
 import { addToDate, getDate, getCurrentDate, subtractToDate } from '../utils/dateManager.js';
 
 moment().format('yyyy-MM-dd');
@@ -63,37 +63,19 @@ export default {
       this.buildChartData();
       this.renderChart(this.chartdata, this.chartoptions);
     },
-    increaseDataset(dateLabels, datasetData, index, nextLabel) {
-      dateLabels.splice(index + 1, 0, nextLabel);
-      datasetData.splice(index + 1, 0, datasetData[index]);
-    },
     getSecondToLastItem(array) {
       const numberTwo = 2;
 
       return array[array.length - numberTwo];
     },
-    fillDatasetGaps(dateLabels, datasetData) {
-      if (dateLabels.length === 0) return;
-      let index = 0;
-      const lastLabel = getDate(
-        subtractToDate(getCurrentDate(), 1, this.dateTypeSelector, { dayOfWeek: this.dayOfWeek }),
-        this.dateTypeSelector,
-        this.dayOfWeek,
-        true
-      );
-      let currentLabel = dateLabels[index];
-      let nextLabel;
-      while (moment(currentLabel).isSameOrBefore(lastLabel, 'day')) {
-        nextLabel = addToDate(currentLabel, 1, this.dateTypeSelector, this.dayOfWeek);
-        if (!dateLabels.includes(nextLabel)) this.increaseDataset(dateLabels, datasetData, index, nextLabel);
-        index++;
-        currentLabel = dateLabels[index];
-      }
-    },
     buildChartData() {
       let dateLabels = getLabels(this.filteredActivities);
       const currentDataset = buildChartDataSet(this.filteredActivities, dateLabels, 'Current Progression', { color: 'black', fill: false });
-      this.fillDatasetGaps(dateLabels, currentDataset.data);
+      fillDatasetGaps(
+        dateLabels,
+        currentDataset.data,
+        { dateTypeSelector: this.dateTypeSelector, dayOfWeek: this.dayOfWeek }
+      );
       const currentProjection = this.projectData(
         this.speed,
         this.timeUnitsForward,
